requirements.txt
psycopg2-binary==2.9.1
tqdm==4.62.3